rename input component
