+++ client/src/components/Input/Input.types.ts
@@ -1,0 +1,5 @@
+import { InputHTMLAttributes } from 'react';
+
+import { ClassName } from 'src/types';
+
+export type InputProps = InputHTMLAttributes<HTMLInputElement> & ClassName;
