--- client/src/components/TextInput/TextInput.module.scss
@@ -1,52 +1,0 @@
-@import "src/styles/palette";
-@import "src/styles/variables";
-
-.textInput {
-  position: relative;
-  width: 100%;
-  display: block;
-
-  span {
-    color: $black;
-  }
-}
-
-.input {
-  position: relative;
-  width: 100%;
-  margin-top: 4px;
-}
-
-.button {
-  width: 32px;
-  height: 24px;
-  padding: 0;
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-  color: $dark-grey;
-  transition: color $transition-duration;
-  position: absolute;
-  right: 12px;
-  top: 8px;
-  outline: none;
-  background-color: $white;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  &:hover,
-  &:focus {
-    color: $black;
-  }
-
-  svg {
-    width: 24px;
-    height: 24px;
-
-    * {
-      fill: currentColor;
-      transition: fill $transition-duration;
-    }
-  }
-}
