--- client/src/components/TextInput/TextInput.types.ts
@@ -1,6 +1,0 @@
-import { TextFieldProps } from '../TextField/TextField.types';
-
-export type TextInputProps = TextFieldProps & {
-  label: string;
-  forPassword?: boolean;
-};
