--- client/src/components/TextInput/index.ts
@@ -1,3 +1,0 @@
-import TextInput from './TextInput';
-
-export default TextInput;
