+++ client/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Button } from './Button';
 export { default as Card } from './Card';
 export { default as H1 } from './H1';
+export { default as Input } from './Input';
 export { default as TextField } from './TextField';
-export { default as TextInput } from './TextInput';
