+++ client/src/views/Registration/Registration.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
-import { Button, Card, H1, TextInput } from 'src/components';
+import { Button, Card, H1, TextField } from 'src/components';
 import { PATHS } from 'src/constants';
 
 import styles from './Registration.module.scss';
@@ -15,20 +15,20 @@ const Registration: FC = () => {
       <H1>{t('registration')}</H1>
 
       <form>
-        <TextInput label={t('username')} placeholder={t('usernamePlaceholder')} className={styles.input} />
+        <TextField label={t('username')} placeholder={t('usernamePlaceholder')} className={styles.input} />
 
         <div className={styles.inputGroup}>
-          <TextInput label={t('firstName')} placeholder={t('firstNamePlaceholder')} className={styles.inputItem} />
-          <TextInput label={t('lastName')} placeholder={t('lastNamePlaceholder')} className={styles.inputItem} />
+          <TextField label={t('firstName')} placeholder={t('firstNamePlaceholder')} className={styles.inputItem} />
+          <TextField label={t('lastName')} placeholder={t('lastNamePlaceholder')} className={styles.inputItem} />
         </div>
 
-        <TextInput
+        <TextField
           forPassword={true}
           label={t('password')}
           placeholder={t('passwordPlaceholder')}
           className={styles.input}
         />
-        <TextInput
+        <TextField
           forPassword={true}
           label={t('confirmPassword')}
           placeholder={t('passwordPlaceholder')}
